+++ src/lib.rs
@@ -2,18 +2,18 @@
 //!
 //! The `__geo_interface__` protocol is implemented by most popular geospatial python modules like `shapely`, `geojson`, `geopandas`, ....
 //!
-//! The main struct of this crate is [`GeometryInterface`]. This the docs there for usage examples.
+//! The main struct of this crate is [`Geometry`]. This the docs there for usage examples.
 //!
 //! ## Features
 //!
-//! As rust types exposed to python may not have generic type parameters, there are multiple implementations of the `GeometryInterface` type based
+//! As rust types exposed to python may not have generic type parameters, there are multiple implementations of the `Geometry` type based
 //! on different types for the coordinate values. The default is `f64`, other types can be enabled using the `f32`, `u8`, `u16`, `u32`, `u64`,
-//! `i8`, `i16`, `i32` and `i64` feature gates. The implementation are then available as `py_geo_interface::wrappers::[datatype]::GeometryInterface`.
-//! The default and probably most common used `f64`-variant is also available as `py_geo_interface::GeometryInterface`.
+//! `i8`, `i16`, `i32` and `i64` feature gates. The implementation are then available as `py_geo_interface::wrappers::[datatype]::Geometry`.
+//! The default and probably most common used `f64`-variant is also available as `py_geo_interface::Geometry`.
 //!
 //! The `wkb` feature adds support for exchanging geometries using the Well-Known-Binary format. The `wkb`-property of `shapely`
-//! geometries will be used when found. Additionally, the `GeometryInterface`-type exposed to python will have a `wkb`-property
-//! itself. WKB is only supported for the `f64`-variant of the `GeometryInterface`, the feature is disabled per default.
+//! geometries will be used when found. Additionally, the `Geometry`-type exposed to python will have a `wkb`-property
+//! itself. WKB is only supported for the `f64`-variant of the `Geometry`, the feature is disabled per default.
 //!
 //! ## Examples
 //!
@@ -21,9 +21,9 @@
 //!
 //! #[include]
 //! ```rust
-//! use geo_types::{Geometry, Point};
+//! use geo_types::{Geometry as GtGeometry, Point};
 //! use pyo3::{prepare_freethreaded_python, Python};
-//! use py_geo_interface::GeometryInterface;
+//! use py_geo_interface::Geometry;
 //!
 //! prepare_freethreaded_python();
 //!
@@ -39,25 +39,25 @@
 //! "#, None, None).unwrap();
 //!
 //!     // create an instance of the class and extract the geometry
-//!     py.eval(r#"Something()"#, None, None)?.extract::<GeometryInterface>()
+//!     py.eval(r#"Something()"#, None, None)?.extract::<Geometry>()
 //! }).unwrap();
-//! assert_eq!(geom.0, Geometry::Point(Point::new(5.0_f64, 3.0_f64)));
+//! assert_eq!(geom.0, GtGeometry::Point(Point::new(5.0_f64, 3.0_f64)));
 //! ```
 //!
 //! ### Pass geometries from Rust to Python:
 //!
 //! ```rust
-//! use geo_types::{Geometry, Point};
+//! use geo_types::{Geometry as GtGeometry, Point};
 //! use pyo3::{prepare_freethreaded_python, Python};
 //! use pyo3::types::{PyDict, PyTuple};
 //! use pyo3::IntoPy;
-//! use py_geo_interface::GeometryInterface;
+//! use py_geo_interface::Geometry;
 //!
 //! prepare_freethreaded_python();
 //!
 //! Python::with_gil(|py| {
 //!
-//!     let geom: GeometryInterface = Point::new(10.6_f64, 23.3_f64).into();
+//!     let geom: Geometry = Point::new(10.6_f64, 23.3_f64).into();
 //!     let mut locals = PyDict::new(py);
 //!     locals.set_item("geom", geom.into_py(py)).unwrap();
 //!
@@ -71,9 +71,6 @@
 pub mod from_py;
 pub mod to_py;
 pub mod wrappers;
-
-#[cfg(feature = "f64")]
-pub mod series;
 
 #[cfg(feature = "wkb")]
 pub mod wkb;
@@ -103,4 +100,8 @@ impl<T: CoordNum + IntoPy<Py<PyAny>> + ExtractFromPyFloat + ExtractFromPyInt + W
 impl<T: CoordNum + IntoPy<Py<PyAny>> + ExtractFromPyFloat + ExtractFromPyInt> PyCoordNum for T {}
 
 #[cfg(feature = "f64")]
-pub use crate::wrappers::f64::GeometryInterface;
+pub use crate::wrappers::f64::Geometry;
+#[cfg(feature = "f64")]
+pub use crate::wrappers::f64::GeometryVec;
+#[cfg(feature = "f64")]
+pub use crate::wrappers::f64::GeometryVecFc;