+++ src/wrappers.rs
@@ -3,15 +3,18 @@ macro_rules! dt_mod {
         pub mod $mod_name {
             use crate::from_py::AsGeometry;
             use crate::to_py::AsGeoInterface;
+            use crate::to_py::AsGeoInterfaceFeatureCollection;
+            use crate::to_py::AsGeoInterfaceList;
+            use geo_types::Geometry as GtGeometry;
             use pyo3::prelude::*;
 
             /// Exchanges vector geometries between Rust and Python using [pyo3](https://pyo3.rs) and [Pythons `__geo_interface__` protocol](https://gist.github.com/sgillies/2217756).
             #[derive(Debug)]
             #[pyclass]
-            pub struct GeometryInterface(pub geo_types::Geometry<$coord_type>);
+            pub struct Geometry(pub GtGeometry<$coord_type>);
 
             #[pymethods]
-            impl GeometryInterface {
+            impl Geometry {
                 #[getter]
                 fn __geo_interface__(&self, py: Python) -> PyResult<PyObject> {
                     self.0.as_geointerface_pyobject(py)
@@ -26,23 +29,23 @@ macro_rules! dt_mod {
                 }
             }
 
-            impl<'source> FromPyObject<'source> for GeometryInterface {
+            impl<'source> FromPyObject<'source> for Geometry {
                 fn extract(ob: &'source PyAny) -> PyResult<Self> {
-                    Ok(GeometryInterface(ob.as_geometry()?))
+                    Ok(Self(ob.as_geometry()?))
                 }
             }
 
-            impl From<geo_types::Geometry<$coord_type>> for GeometryInterface {
-                fn from(geom: geo_types::Geometry<$coord_type>) -> Self {
+            impl From<GtGeometry<$coord_type>> for Geometry {
+                fn from(geom: GtGeometry<$coord_type>) -> Self {
                     Self(geom)
                 }
             }
 
             macro_rules! geometry_enum_from_impl {
                 ($geom_type:ty, $enum_variant_name:ident) => {
-                    impl From<$geom_type> for GeometryInterface {
+                    impl From<$geom_type> for Geometry {
                         fn from(g: $geom_type) -> Self {
-                            GeometryInterface(geo_types::Geometry::$enum_variant_name(g))
+                            Geometry(GtGeometry::$enum_variant_name(g))
                         }
                     }
                 };
@@ -61,9 +64,97 @@ macro_rules! dt_mod {
             geometry_enum_from_impl!(geo_types::Line<$coord_type>, Line);
             geometry_enum_from_impl!(geo_types::Triangle<$coord_type>, Triangle);
 
-            impl From<GeometryInterface> for geo_types::Geometry<$coord_type> {
-                fn from(gw: GeometryInterface) -> Self {
+            impl From<Geometry> for GtGeometry<$coord_type> {
+                fn from(gw: Geometry) -> Self {
                     gw.0
+                }
+            }
+
+            /// Vec of geometries
+            ///
+            /// Accessible from python via `__geo_interface__` as a list of geometries.
+            #[derive(Debug)]
+            #[pyclass]
+            pub struct GeometryVec(pub Vec<GtGeometry<$coord_type>>);
+
+            #[pymethods]
+            impl GeometryVec {
+                #[getter]
+                fn __geo_interface__(&self, py: Python) -> PyResult<PyObject> {
+                    self.0.as_geointerface_list_pyobject(py)
+                }
+            }
+
+            impl<'source> FromPyObject<'source> for GeometryVec {
+                fn extract(ob: &'source PyAny) -> PyResult<Self> {
+                    Ok(ob.as_geometry_vec()?)
+                }
+            }
+
+            impl From<GeometryVec> for Vec<GtGeometry<$coord_type>> {
+                fn from(gv: GeometryVec) -> Self {
+                    gv.0
+                }
+            }
+
+            impl From<Vec<GtGeometry<$coord_type>>> for GeometryVec {
+                fn from(geoms: Vec<GtGeometry<$coord_type>>) -> Self {
+                    Self(geoms)
+                }
+            }
+
+            pub trait AsGeometryVec {
+                /// Creates a `GeometryVec` from `self`
+                fn as_geometry_vec(&self) -> PyResult<GeometryVec>;
+            }
+
+            impl AsGeometryVec for PyAny {
+                fn as_geometry_vec(&self) -> PyResult<GeometryVec> {
+                    GeometryVec::extract(self)
+                }
+            }
+
+            /// Vec of geometries
+            ///
+            /// Accessible from python via `__geo_interface__` as a FeatureCollection.
+            #[derive(Debug)]
+            #[pyclass]
+            pub struct GeometryVecFc(pub Vec<GtGeometry<$coord_type>>);
+
+            #[pymethods]
+            impl GeometryVecFc {
+                #[getter]
+                fn __geo_interface__(&self, py: Python) -> PyResult<PyObject> {
+                    self.0.as_geointerface_featurecollection_pyobject(py)
+                }
+            }
+
+            impl<'source> FromPyObject<'source> for GeometryVecFc {
+                fn extract(ob: &'source PyAny) -> PyResult<Self> {
+                    Ok(ob.as_geometry_vec_fc()?)
+                }
+            }
+
+            impl From<GeometryVecFc> for Vec<GtGeometry<$coord_type>> {
+                fn from(gv: GeometryVecFc) -> Self {
+                    gv.0
+                }
+            }
+
+            impl From<Vec<GtGeometry<$coord_type>>> for GeometryVecFc {
+                fn from(geoms: Vec<GtGeometry<$coord_type>>) -> Self {
+                    Self(geoms)
+                }
+            }
+
+            pub trait AsGeometryVecFc {
+                /// Creates a `GeometryVecFc` from `self`
+                fn as_geometry_vec_fc(&self) -> PyResult<GeometryVecFc>;
+            }
+
+            impl AsGeometryVecFc for PyAny {
+                fn as_geometry_vec_fc(&self) -> PyResult<GeometryVecFc> {
+                    GeometryVecFc::extract(self)
                 }
             }
         }